requirements.txt
flask
flask-smorest
python-dotenv
requests
flask-jwt-extended
passlib
gunicorn
flask-injector
boto3